Fix issue #386
Moved two functions into [data/ContentProviderUtils.java] and deleted [util/TrackUtils.java].
