+++ src/main/java/de/dennisguse/opentracks/publicapi/StartRecording.java
@@ -9,7 +9,6 @@ import de.dennisguse.opentracks.data.models.Track;
 import de.dennisguse.opentracks.services.TrackRecordingService;
 import de.dennisguse.opentracks.settings.PreferencesUtils;
 import de.dennisguse.opentracks.util.IntentDashboardUtils;
-import de.dennisguse.opentracks.util.TrackUtils;
 
 public class StartRecording extends AbstractAPIActivity {
 
@@ -41,7 +40,7 @@ public class StartRecording extends AbstractAPIActivity {
         ContentProviderUtils contentProviderUtils = new ContentProviderUtils(this);
         Track track = contentProviderUtils.getTrack(trackId);
 
-        TrackUtils.updateTrack(this, track,
+        ContentProviderUtils.updateTrack(this, track,
                 bundle.getString(EXTRA_TRACK_NAME, null),
                 bundle.getString(EXTRA_TRACK_CATEGORY, null),
                 bundle.getString(EXTRA_TRACK_ICON, null),
